site-react: fix home 404-on-hydration — split into server page + HomeClient
Visual fidelity pass caught it: the home rendered Next's not-found after
hydration because app/page.tsx was a top-level "use client" page, which 404s
under output:export. Refactored to the server-page + client-child pattern
every other route already uses (server page.tsx with metadata → HomeClient).
Verified via screenshot: full home renders. All other page.tsx confirmed
server components.
